pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "fpdf>=1.7.2",
    "google-api-python-client>=2.173.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth>=2.40.3",
    "google-auth-oauthlib>=1.2.2",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-telegram-bot>=22.1",
    "requests>=2.32.4",
    "telegram>=0.0.1",
    "fpdf2>=2.8.3",
]
